+++ src/main/java/com/qsmy/springboot/repository/UserRepository.java
@@ -3,5 +3,8 @@ package com.qsmy.springboot.repository;
 import com.qsmy.springboot.bean.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+/**
+ * @author qsmy
+ */
+public interface UserRepository extends JpaRepository<User, Long> {
 }
